tests/wpt/tests/css/css-text/parsing/word-boundary-expansion-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing word-boundary-expansion with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-text-4/#word-boundary-expansion">
<meta name="assert" content="word-boundary-expansion supports only the grammar 'none | space | ideographic-space'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("word-boundary-expansion", "auto");
test_invalid_value("word-boundary-expansion", "normal");
test_invalid_value("word-boundary-expansion", "spaces");
test_invalid_value("word-boundary-expansion", "space ideographic-space");
</script>
</body>
</html>
